+++ app/initializers/main.js
@@ -1,4 +1,5 @@
 import { registerDeprecationHandler } from '@ember/debug';
+import debug from 'debug';
 
 /**
  * Main Initialzer
@@ -7,10 +8,11 @@ import { registerDeprecationHandler } from '@ember/debug';
  * @class Initializer.Main
  */
 export function initialize(/* application */) {
-  // application.inject('route', 'foo', 'service:foo');
+
+  let log = debug('initializer:main');
 
   // Muestra un aviso.
-  console.warn('The deprecations are hidden but are showing in Ember\'s Inspector deprecations tab. See main initializer for more information.');
+  log('The deprecations are hidden but are showing in Ember\'s Inspector deprecations tab. See main initializer for more information.');
 
   /**
    * Deprecation Handler. Oculta las depreciaciones. See this: https://guides.emberjs.com/v3.4.0/configuring-ember/handling-deprecations/.