+++ app/routes/application.js
@@ -1,4 +1,3 @@
-// import injectScript from 'ember-inject-script';
 import config from '../config/environment';
 import EmberObject from '@ember/object';
 import Route from '@ember/routing/route';
@@ -8,8 +7,6 @@ import { hash } from 'rsvp';
 import { inject as service } from '@ember/service';
 import { isBlank } from '@ember/utils';
 import { Promise } from 'rsvp';
-import { set } from '@ember/object';
-import debugLogger from 'ember-debug-logger';
 
 /**
  * Application Route
@@ -17,12 +14,6 @@ import debugLogger from 'ember-debug-logger';
  * @class Route.Application
  */
 export default Route.extend({
-
-  debug: debugLogger(),
-
-  activate() {
-    this.debug('Hello from the application index.');
-  },
 
   /**
    * Spreadsheets Service
@@ -49,40 +40,33 @@ export default Route.extend({
   ajax: service(),
 
   /**
-   * TODO: Hacer esto en un lugar más decente, por amor al Señor
-   */
-  /**
-   * Before model hook. Setear la URL de datos y de configuración en el servicio spreadsheet. Además procesar los campos de información general del perfil.
+   * Establecer la 'URL' de los datos y configuraciones en el servicio spreadsheet. Además procesar los campos de serialización.
    *
    * @method beforeModel
-   * @return {Object}            Un objecto con los datos de la configuración del proyecto desde el sevicio Spreadsheets.
    */
   beforeModel() {
     const spreadsheetService = this.get('spreadsheets');
 
-    // TODO: Agregar validación: si config.APP.dataSpreadsheetSourceUrl no esta definida,
-    // lanzar error
-
     return this.get('ajax')
-
+      // Obtiene dataSpreadsheetSourceUrl de las configuraciones
       .request(config.APP.dataSpreadsheetSourceUrl, { dataType: 'text' })
-
       .then((response) => {
+        // Agrega la url de datos al servicio
         spreadsheetService.set('dataSpreadsheetUrl', response);
+        // Utiliza la misma url de datos para configuraciones
         spreadsheetService.set('configSpreadsheetUrl', response);
-
-        // Si config.APP.configSpreadsheetSourceUrl está definida, entonces obtener
-        // también ese valor y setearlo en el spreadsheet service
+        // En el caso que un url para configuraciones existe la obtiene y la agrega
         if (!isBlank(config.APP.configSpreadsheetSourceUrl)) {
           return this.get('ajax')
             .request(config.APP.configSpreadsheetSourceUrl, { dataType: 'text' })
             .then((response) => spreadsheetService.set('configSpreadsheetUrl', response));
         }
-
         return Promise.resolve(this);
       })
 
+      // Obtiene datos de configuraciones para serializar
       .then(() => RSVP.all([
+
         /**
          * Setear la información general del perfil mediante la parametrización
          * proveniente de la configuración
@@ -108,42 +92,21 @@ export default Route.extend({
          * Setear la información de recuadros del perfil mediante la parametrización
          * proveniente de la configuración
          */
-        // spreadsheetService
-        //   .fetchConfig('perfil-recuadros-configuracion')
-        //   .then((configuracionData) => {
-        //     let perfilRecuadrosDataArray = A([]);
-
-        //     A(configuracionData).forEach((item) => {
-        //       perfilRecuadrosDataArray.pushObject({
-        //         field: item.field,
-        //         label: item.label
-        //       });
-        //     });
-
-        //     let prefilSerializer = this.store.serializerFor('magistrate');
-
-        //     prefilSerializer.set('recuadrosFields', perfilRecuadrosDataArray);
-        //   }),
-
-        // Información general de diputado
-        // TODO: Evaluar pertinencia, ya que es una funcionalidad específica de
-        // Elección PDH
         spreadsheetService
-          .fetchConfig('diputado-informacion-general-configuracion')
+          .fetchConfig('perfil-recuadros-configuracion')
           .then((configuracionData) => {
-            let diputadoDataArray = A([]);
+            let perfilRecuadrosDataArray = A([]);
 
             A(configuracionData).forEach((item) => {
-              diputadoDataArray.pushObject({
+              perfilRecuadrosDataArray.pushObject({
                 field: item.field,
                 label: item.label
               });
             });
 
-            let diputadoSerializer = this.store.serializerFor('diputado-comision');
+            let prefilSerializer = this.store.serializerFor('magistrate');
 
-            diputadoSerializer.set('informacionGeneralFields', A());
-            diputadoSerializer.set('frenteAFrenteFields', A());
+            prefilSerializer.set('recuadrosFields', perfilRecuadrosDataArray);
           }),
 
         /**
@@ -170,17 +133,16 @@ export default Route.extend({
   },
 
   /**
-   * Model hook
+   * Datos principales de la aplicación.
    *
    * @method model
-   * @return {Object} Datos de nuestro aplicación, perfiles y datos de configuración.
+   * @return {Object} Perfiles, config, navbarLinks.
    */
   model() {
     const spreadsheet = this.get('spreadsheets');
     const _routing = this.get('_routing');
 
     return hash({
-      // partidos: this.store.findAll('partido'),
       perfiles: this.store.findAll('magistrate'),
       config: spreadsheet.fetchConfig('configuracion')
         .then((configuracion) => {
@@ -190,74 +152,15 @@ export default Route.extend({
             configObject.set(item.key, item.value);
           });
 
-          /**
-           * Inject HelloBar if defined
-           */
-          // if (!isBlank(configObject.helloBarUrl)) {
-          //   injectScript(configObject.helloBarUrl);
-          // }
-
           return configObject;
         }),
-
-      /**
-       * Header links, top right
-       */
-      navbarLinks: spreadsheet.fetchConfig('navbar-links').then((links) => {
-        return A(links).filter((link) => {
-          return _routing.hasRoute(link.route);
-        });
-      }),
-
-      /**
-       * Front page image links.
-       *
-       * If the row does not include a link property it gets dissmissed
-       */
-      mainPageLinks: spreadsheet.fetchConfig('main-page-links').then((links) => {
-        return A(links).filter((link) => {
-          if (link.link) {
-            return true;
-          }
-
-          return _routing.hasRoute(link.route);
-        });
-      }),
-
-      /**
-       * Main page slider profiles list
-       */
-      mainPageSliderData: spreadsheet.fetchConfig('main-page-slider-data'),
-
-      institucionData: spreadsheet
-        .fetch('institucion-data')
-        .then((institucionData) => {
-          let institucionDataObject = EmberObject.create();
-
-          A(institucionData).forEach((item) => {
-            institucionDataObject.set(item.key, item.value);
+      navbarLinks: spreadsheet.fetchConfig('navbar-links')
+        .then((links) => {
+          return A(links).filter((link) => {
+            return _routing.hasRoute(link.route);
           });
-
-          return institucionDataObject;
         }),
-
-      frontTableVisualizationConfig: spreadsheet.fetchConfig('front-table-visualization-config')
     });
-  },
-
-  /**
-   * Levanta un controlador y asigna unos valores.
-   *
-   * @method setupController
-   * @param  {[type]} controller Clase controller.
-   * @param  {[type]} model      Modelo de esta ruta.
-   */
-  setupController(controller, model) {
-    this._super(controller, model);
-
-    set(model.config, 'navbarLinks', model.navbarLinks);
-    set(model.config, 'mainPageLinks', model.mainPageLinks);
-    set(model.config, 'mainPageSliderData', model.mainPageSliderData);
   },
 
   /**
@@ -266,6 +169,7 @@ export default Route.extend({
    * @type {Object}
    */
   actions: {
+    // Utilizado para seleccionar un perfil en la caja de busqueda.
     selectPerfil(candidato) {
       this.transitionTo('perfil', candidato.get('id'));
     }